distance model

+++ app/src/main/java/kr/yhs/traffic/models/StationInfo.kt
@@ -4,8 +4,8 @@ package kr.yhs.traffic.models
 data class StationInfo(
     val name: String,
     val id: Int,
-    val posX: Float,
-    val posY: Float,
+    val posX: Double,
+    val posY: Double,
     val displayId: Any?,
     val stationId: Int,
     val type: Int
